fix: AccountPageClient 분리 (#21)
